requirements.txt
alabaster==0.7.12
asgiref==3.2.7
beautifulsoup4==4.9.0
bkcharts==0.2
bleach==3.1.0
bootstrap4==0.1.0
boto==2.49.0
Bottleneck==1.2.1
certifi==2020.4.5.1
coverage==5.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.6
django-bootstrap4==1.1.1
django-chartjs==2.1.0
django-jquery==3.1.0
django-leaflet==0.26.0
django-stdimage==5.1.1
gunicorn==20.0.4
Pillow==7.1.2
progressbar2==3.51.3
psycopg2-binary==2.8.5
python-utils==2.4.0
pytz==2020.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
whitenoise==5.0.1
wincertstore==0.2
